+++ Assets/Scripts/Player/PlayerStatsGUI.cs
@@ -7,7 +7,7 @@ public class PlayerStatsGUI : MonoBehaviour
 {
     [SerializeField] public Slider slider;
 
-    public void ShowHealth(int health)
+    public void ShowHealth(float health)
     {
         slider.value = health;
     }
